VotingApplication/VotingApplication.Web.Api/Controllers/API Controllers/VoteController.cs
using System.Collections.Generic;
using System.Data.Entity;
using System.Linq;
using System.Net;
using System.Net.Http;
using VotingApplication.Data.Context;
using VotingApplication.Data.Model;

namespace VotingApplication.Web.Api.Controllers
{
    public class VoteController : WebApiController
    {
        public VoteController() : base() { }
        public VoteController(IContextFactory contextFactory) : base(contextFactory) { }

        #region Get
        public override HttpResponseMessage Get()
        {
            using(var context = _contextFactory.CreateContext())
            {
                return this.Request.CreateResponse(HttpStatusCode.OK, context.Votes.Include(v => v.Option).Include(v => v.User).ToList<Vote>());
            }
        }

        public override HttpResponseMessage Get(long id)
        {
            using (var context = _contextFactory.CreateContext())
            {
                Vote voteForId = context.Votes.Where(v => v.Id == id).Include(v => v.Option).Include(v => v.User).FirstOrDefault();
                if (voteForId == null)
                {
                    return this.Request.CreateErrorResponse(HttpStatusCode.NotFound, string.Format("Vote {0} not found", id));
                }
                else
                {
                    return this.Request.CreateResponse(HttpStatusCode.OK, voteForId);
                }
            }
        } 
        #endregion

        #region PUT

        public virtual HttpResponseMessage Put(Vote newVote)
        {
            return this.Request.CreateErrorResponse(HttpStatusCode.MethodNotAllowed, "Cannot use PUT on this controller");
        }

        #endregion
    }
}